refactor folder and files\
