--- src/components/Analytics/components/ContentVelocity/index.js
@@ -1,1 +1,0 @@
-export { ContentVelocity } from "./ContentVelocity";
