--- src/components/Analytics/components/RecentlyEdited/index.js
@@ -1,1 +1,0 @@
-export { RecentlyEdited } from "./RecentlyEdited";
