+++ src/components/Analytics/graph/SocialTraffic/SocialTraffic.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import { Doughnut } from "react-chartjs-2";
-import GraphContainer from '../GraphContainer';
+import GraphContainer from '../../../ui/GraphContainer';
 
 export const SocialTraffic = ({ setGALegacyStatus, instanceZUID, profileID, data, domainSet }) => {
