+++ src/components/Analytics/graph/index.js
@@ -1,0 +1,11 @@
+import { PageviewTraffic } from "./PageviewTraffic";
+import { InboundTraffic } from "./InboundTraffic";
+import { SocialTraffic } from "./SocialTraffic";
+import { TopPerforming } from "./TopPerforming";
+
+export {
+    PageviewTraffic,
+    InboundTraffic,
+    SocialTraffic,
+    TopPerforming
+}
